cfb_server/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-environ==0.13.0
django-taggit==6.1.0
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
psycopg2==2.9.11
pymemcache==4.0.0
python-memcached==1.62
sqlparse==0.5.5
